requirements/base.txt
Django==5.2.9
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.28.0
psycopg[binary]==3.2.9
redis==5.0.8
python-dotenv==1.0.1
celery==5.4.0
opensearch-py==2.6.0

# ingestion deps
requests==2.32.3
beautifulsoup4==4.12.3
pypdf==5.1.0
python-docx==1.1.2
Pillow==10.4.0
